Remove old cli file
