+++ package.json
@@ -9,7 +9,7 @@
   "bin": "scripts/cli.js",
   "scripts": {
     "test": "eslint .",
-    "create": "./scripts/cli2.0.js"
+    "create": "./scripts/cli.js"
   },
   "dependencies": {
     "@types/mongoose": "^5.7.7",
@@ -18,8 +18,7 @@
     "eris": "^0.12.0",
     "fs-extra": "^9.0.0",
     "inquirer": "^7.1.0",
-    "minimist": "^1.2.5",
-    "shelljs": "^0.8.3"
+    "minimist": "^1.2.5"
   },
   "devDependencies": {
     "@axonteam/eslint-config": "^2.2.1",
